+++ src/models/command/lampSetLightCommand.ts
@@ -11,6 +11,6 @@ export class LampSetLightCommand extends ActuatorSetStateCommand {
   }
 
   public override get logMessage(): string {
-    return `Lamp setLight to ${this.on} for reason: ${this.reasonTrace}`;
+    return `Lamp setLight to ${this.on} with timeout ${this.timeout} for reason: ${this.reasonTrace}`;
   }
 }