+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Improve IoBrokerDevice set State Error Logging
+* (theimo1221) Improve Lamp/Dimmer/WLed Command Logging
 
 ## 3.0.0-alpha.16 (2024-03-18)